+++ x360ce.Engine/Common/XInputMaskScannerEventArgs.cs
@@ -17,10 +17,10 @@ namespace x360ce.Engine
 		public FileInfo GameFileInfo { get; set; }
 
 		public List<DirectoryInfo> Directories;
-		public int DirectoryIndex { get; set; }
+		public long DirectoryIndex { get; set; }
 
 		public List<FileInfo> Files;
-		public int FileIndex { get; set; }
+		public long FileIndex { get; set; }
 
 		public int Skipped { get; set; }
